Updated search services for use with search-api
